refactor:  night study feature to management
Renames the night study feature to management night study to better reflect its purpose.

Updates related navigation, dependency injection, and package structures across both student and teacher applications.
