+++ dodam-student/src/main/kotlin/com/b1nd/dodam/student/util/DodamApplication.kt
@@ -48,7 +48,7 @@ import com.b1nd.dodam.network.outing.di.outingDataSourceModule
 import com.b1nd.dodam.network.point.di.pointDataSourceModule
 import com.b1nd.dodam.network.schedule.di.scheduleDatasourceModule
 import com.b1nd.dodam.network.upload.di.uploadDatasourceModule
-import com.b1nd.dodam.nightstudy.di.nightStudyViewModelModule
+import com.b1nd.dodam.managementnightstudy.di.managementNightStudyViewModelModule
 import com.b1nd.dodam.notice.di.noticeViewModelModule
 import com.b1nd.dodam.outing.di.outingViewModelModule
 import com.b1nd.dodam.parent.childrenmanage.di.childrenManageViewModelModule
@@ -114,7 +114,7 @@ class DodamApplication : Application() {
                     loginViewModelModule,
                     registerViewModelModule,
                     settingViewModelModule,
-                    nightStudyViewModelModule,
+                    managementNightStudyViewModelModule,
                     outingViewModelModule,
                     bundleIdInfoRepositoryModule,
                     bundleIdInfoDataSourceModule,
@@ -139,7 +139,8 @@ class DodamApplication : Application() {
                     divisionDataSourceModule,
                     noticeRepositoryModule,
                     noticeDatasourceModule,
-                    approveNightStudyViewModelModule
+                    approveNightStudyViewModelModule,
+                    managementNightStudyViewModelModule,
                 ) + mainViewModelModules,
             )
         }
