+++ dodam-teacher-android/src/commonMain/kotlin/com/b1nd/dodam/teacher/Koin.kt
@@ -42,7 +42,7 @@ import com.b1nd.dodam.network.outing.di.outingDataSourceModule
 import com.b1nd.dodam.network.point.di.pointDataSourceModule
 import com.b1nd.dodam.network.schedule.di.scheduleDatasourceModule
 import com.b1nd.dodam.network.upload.di.uploadDatasourceModule
-import com.b1nd.dodam.nightstudy.di.nightStudyViewModelModule
+import com.b1nd.dodam.managementnightstudy.di.managementNightStudyViewModelModule
 import com.b1nd.dodam.notice.di.noticeViewModelModule
 import com.b1nd.dodam.noticecreate.di.noticeCreateViewModelModule
 import com.b1nd.dodam.outing.di.outingViewModelModule
@@ -83,7 +83,7 @@ fun initKoin(block: KoinApplication.() -> Unit = {}) {
             scheduleDatasourceModule,
             DodamTeacherAppViewModelModule,
             mealViewModelModule,
-            nightStudyViewModelModule,
+            managementNightStudyViewModelModule,
             memberRepositoryModule,
             memberDataSourceModule,
             pointViewModelModule,
