+++ feature/management-nightstudy/src/androidMain/kotlin/com/b1nd/dodam/nightstudy/NightStudyScreenPreview.kt
@@ -2,6 +2,7 @@ package com.b1nd.dodam.nightstudy
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import com.b1nd.dodam.managementnightstudy.NightStudyScreen
 
 @Preview
 @Composable
