+++ feature/management-nightstudy/src/commonMain/kotlin/com/b1nd/dodam/managementnightstudy/di/ViewModelModule.kt
@@ -1,9 +1,9 @@
-package com.b1nd.dodam.nightstudy.di
+package com.b1nd.dodam.managementnightstudy.di
 
-import com.b1nd.dodam.nightstudy.viewmodel.NightStudyViewModel
+import com.b1nd.dodam.managementnightstudy.viewmodel.NightStudyViewModel
 import org.koin.compose.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
-val nightStudyViewModelModule = module {
+val managementNightStudyViewModelModule = module {
     viewModel { NightStudyViewModel() }
 }
