+++ feature/management-nightstudy/src/commonMain/kotlin/com/b1nd/dodam/managementnightstudy/navigation/NightStudyNavigation.kt
@@ -1,4 +1,4 @@
-package com.b1nd.dodam.nightstudy.navigation
+package com.b1nd.dodam.managementnightstudy.navigation
 
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
@@ -7,18 +7,18 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.b1nd.dodam.nightstudy.NightStudyScreen
+import com.b1nd.dodam.managementnightstudy.NightStudyScreen
 
 const val NIGHT_STUDY_ROUTE = "night"
 
-fun NavController.navigateToNightStudy(
+fun NavController.navigateToManagementNightStudy(
     navOptions: NavOptions? = androidx.navigation.navOptions {
         launchSingleTop = true
     },
 ) = navigate(NIGHT_STUDY_ROUTE, navOptions)
 
 @ExperimentalMaterial3Api
-fun NavGraphBuilder.nightStudyScreen(navigateToApproveStudy: () -> Unit) {
+fun NavGraphBuilder.managementNightStudyScreen(navigateToApproveStudy: () -> Unit) {
     composable(
         route = NIGHT_STUDY_ROUTE,
         enterTransition = { EnterTransition.None },
