+++ feature/management-nightstudy/src/commonMain/kotlin/com/b1nd/dodam/managementnightstudy/state/NightStudyUiState.kt
@@ -1,4 +1,4 @@
-package com.b1nd.dodam.nightstudy.state
+package com.b1nd.dodam.managementnightstudy.state
 
 import com.b1nd.dodam.data.nightstudy.model.NightStudy
 import kotlinx.collections.immutable.ImmutableList
@@ -10,7 +10,8 @@ data class NightStudyScreenUiState(
 )
 
 sealed interface NightStudyUiState {
-    data class Success(val pendingCnt: Int, val ingData: ImmutableList<NightStudy>) : NightStudyUiState
+    data class Success(val pendingCnt: Int, val ingData: ImmutableList<NightStudy>) :
+        NightStudyUiState
     data object Loading : NightStudyUiState
     data object Error : NightStudyUiState
 }
